backend/requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
openai
python-multipart
pydantic
pytube
youtube-transcript-api